# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Homegrown Hoops Basketball Stats Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/pending-accounts:
    get:
      operationId: listPendingAccounts
      tags: [admin]
      summary: List accounts awaiting approval (parent/manager signups)
      responses:
        "200":
          description: List of pending accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingAccountItem"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/pending-accounts/{clerkUserId}/approve:
    post:
      operationId: approvePendingAccount
      tags: [admin]
      summary: Approve a pending account (sets role to requestedRole, clears isPending)
      parameters:
        - name: clerkUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingAccountItem"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Account not found
  /admin/pending-accounts/{clerkUserId}/reject:
    post:
      operationId: rejectPendingAccount
      tags: [admin]
      summary: Reject a pending account (clears isPending, keeps as player)
      parameters:
        - name: clerkUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingAccountItem"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Account not found
...
  /manager/my-teams:
    get:
      operationId: getManagerMyTeams
      tags: [manager]
      summary: List teams managed by the current user (admin sees all)
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "401":
          description: Unauthorized
        "403":
          description: Manager access required
  /manager/teams:
    post:
      operationId: createManagerTeam
      tags: [manager]
      summary: Create a new team as manager (auto-approved, no admin review needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagerCreateTeamBody"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Manager access required
  /manager/delegations:
    get:
      operationId: listManagerDelegations
      tags: [manager]
      summary: List active (unused) delegations for the current manager
      responses:
        "200":
          description: List of delegations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagerDelegation"
        "401":
          description: Unauthorized
        "403":
          description: Manager access required
    post:
      operationId: createManagerDelegation
      tags: [manager]
      summary: Grant one-game tracking access for a team to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDelegationBody"
      responses:
        "201":
          description: Delegation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerDelegation"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Manager access required or team not managed
        "404":
          description: Delegatee not found
  /manager/delegations/{id}:
    delete:
      operationId: deleteManagerDelegation
      tags: [manager]
      summary: Remove a delegation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Manager access required
        "404":
          description: Delegation not found
...
components:
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        abbreviation:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: string
        secondaryColor:
          type: string
        league:
          type: ["string", "null"]
        managerClerkUserId:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - city
        - abbreviation
        - wins
        - losses
        - primaryColor
        - secondaryColor
        - createdAt
...
    PendingAccountItem:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        requestedRole:
          type: ["string", "null"]
        requestedTeamInfo:
          type: ["object", "null"]
          description: "Team details submitted during manager signup"
          properties:
            teamName:
              type: string
            league:
              type: ["string", "null"]
            city:
              type: ["string", "null"]
            roster:
              type: array
              items:
                type: object
                properties:
                  jerseyNumber:
                    type: string
                  playerName:
                    type: string
        createdAt:
          type: string
      required:
        - id
        - clerkUserId
        - firstName
        - lastName
        - createdAt
...
    ManagerCreateTeamBody:
      type: object
      properties:
        name:
          type: string
        city:
          type: ["string", "null"]
        league:
          type: ["string", "null"]
      required:
        - name
    ManagerDelegation:
      type: object
      properties:
        id:
          type: integer
        managerClerkUserId:
          type: string
        delegateeClerkUserId:
          type: string
        teamId:
          type: integer
        used:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - managerClerkUserId
        - delegateeClerkUserId
        - teamId
        - used
        - createdAt
    CreateDelegationBody:
      type: object
      properties:
        delegateeClerkUserId:
          type: string
        teamId:
          type: integer
      required:
        - delegateeClerkUserId
        - teamId