# lib/api-zod/src/generated/types/createDelegationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Homegrown Hoops Basketball Stats Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDelegationBody {
  delegateeClerkUserId: string;
  teamId: number;
}