# lib/api-zod/src/generated/types/managerCreateTeamBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Homegrown Hoops Basketball Stats Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface ManagerCreateTeamBody {
  name: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  league?: string | null;
}