# lib/api-zod/src/generated/types/managerDelegation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Homegrown Hoops Basketball Stats Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface ManagerDelegation {
  id: number;
  managerClerkUserId: string;
  delegateeClerkUserId: string;
  teamId: number;
  used: boolean;
  createdAt: string;
}