# lib/api-zod/src/generated/types/pendingAccountItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Homegrown Hoops Basketball Stats Tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { PendingAccountItemRequestedTeamInfo } from "./pendingAccountItemRequestedTeamInfo";

export interface PendingAccountItem {
  id: number;
  clerkUserId: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  requestedRole?: string | null;
  /**
   * Team details submitted during manager signup
   * @nullable
   */
  requestedTeamInfo?: PendingAccountItemRequestedTeamInfo;
  createdAt: string;
}