# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Homegrown Hoops Basketball Stats Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  id: number;
  name: string;
  city: string;
  abbreviation: string;
  wins: number;
  losses: number;
  /** @nullable */
  logoUrl?: string | null;
  primaryColor: string;
  secondaryColor: string;
  /** @nullable */
  league?: string | null;
  /** @nullable */
  managerClerkUserId?: string | null;
  createdAt: string;
}